worked on multiple corrections

+++ src/main/java/com/capstone/furever/controllers/PostController.java
@@ -164,11 +164,13 @@ public class PostController {
 
 
     @PostMapping("/posts/{n}/delete")
-
     public String deletePost(@PathVariable long n) {
         User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        Post post = postDao.findById(n).get();
-        if (user.getId() == post.getUser().getId()) {
+        Post post = postDao.findById(n).orElse(null);
+        if (post != null && user.getId() == post.getUser().getId()) {
+            // delete all associated comments
+            ((PostRepository) postDao).deleteCommentsByPostId(n);
+            // delete the post record
             postDao.deleteById(n);
         }
         return "redirect:/posts";
